refactored SaveScreenshot command to have execute method
